# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InnoDD API",
    "version": "2.0.0",
    "description": "Complete InnoDD drug-discovery platform surface, generated from source (not the Postman collection).\n\n**Two surfaces:**\n- `/api/v1/*` — legacy internal pipeline modules (TxKG, Literature Mining, Drug Curation, Screening, Novelty Search) plus Auth/Users/Items scaffolding and job Orchestration. No X-Drp-Token needed.\n- `/v1/*` — the DRP frontend-facing contract (async job pattern: POST creates a job -> 202 {jobId}, then poll status/result). Needs `X-Drp-Token` (or `Authorization: Bearer`) from `/v1/auth/login`.\n\n**Known gaps / stubs (do not rely on for testing):**\n- `POST /api/v1/auth/login` and `/api/v1/auth/token` are unimplemented stubs (`pass`) in the legacy scaffold.\n- `/api/v1/users/*` imports a `UserRead` model that does not exist in `app/schemas/user.py` — this router likely fails to import and is NOT actually mounted.\n- `/api/v1/agentic/*` never mounts — `agent_routes.py` does not exist as a file; `_include()` swallows the ImportError.\n- Several module routers are double-mounted: Literature Mining, Drug Curation, Screening, and Novelty Search each exist BOTH under their versioned prefix (e.g. `/api/v1/literature-mining/...`) AND again with no prefix as a legacy alias (e.g. `/api/v1/search-keywords/`) — both paths are live simultaneously and hit the same handler.",
    "contact": { "name": "InnoDD API" }
  },
  "servers": [
...
  ],
  "components": {
    "securitySchemes": {
      "drpToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Drp-Token",
        "description": "Access token returned by POST /v1/auth/login or /v1/auth/refresh. The server also accepts a standard Authorization: Bearer header as a fallback."
      }
    },
    "schemas": {
      "JobAccepted": { "type": "object", "properties": { "jobId": { "type": "string" } } },
      "SuccessResponse": { "type": "object", "properties": { "success": { "type": "boolean" } } },
      "UserCreate": { "type": "object", "properties": { "username": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string" } } },
      "UserOut": { "type": "object", "properties": { "id": { "type": "integer" }, "username": { "type": "string" }, "email": { "type": "string" } } },
...
      "CandidateTargetRef": {
        "type": "object",
        "properties": { "name": { "type": "string" }, "uniprotId": { "type": "string", "nullable": true }, "score": { "type": "number", "nullable": true }, "category": { "type": "string", "nullable": true } }
      },
      "ProfileRequest": {
        "type": "object",
        "properties": {
          "target": { "type": "string", "nullable": true },
          "targets": { "type": "array", "items": { "type": "string" } },
          "disease": { "type": "string", "nullable": true },
          "candidateTargets": { "type": "array", "items": { "$ref": "#/components/schemas/CandidateTargetRef" } },
          "weights": { "type": "object", "additionalProperties": { "type": "number" } },
          "skipDailyMed": { "type": "boolean", "default": false },
          "sessionId": { "type": "string", "nullable": true }
        }
      },
...
      "LoginRequest": { "type": "object", "required": ["email", "password"], "properties": { "email": { "type": "string" }, "password": { "type": "string" } } },
      "DrpUser": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "role": { "type": "string" }, "avatarUrl": { "type": "string", "nullable": true }, "email": { "type": "string" } } },
      "LoginResponse": {
        "type": "object",
        "properties": { "accessToken": { "type": "string" }, "refreshToken": { "type": "string" }, "expiresIn": { "type": "integer", "example": 3600 }, "user": { "$ref": "#/components/schemas/DrpUser" } }
      },
      "RefreshRequest": { "type": "object", "required": ["refreshToken"], "properties": { "refreshToken": { "type": "string" } } },
      "RefreshResponse": { "type": "object", "properties": { "accessToken": { "type": "string" }, "expiresIn": { "type": "integer" } } },
      "ForgotPasswordRequest": { "type": "object", "properties": { "email": { "type": "string", "nullable": true } } },
      "OnboardingStatus": { "type": "object", "properties": { "completed": { "type": "boolean" }, "currentStep": { "type": "string" } } },
      "ResearchFocusRequest": { "type": "object", "properties": { "therapeuticAreas": { "type": "array", "items": { "type": "string" } } } },
      "DashboardSummary": { "type": "object", "properties": { "activeProjects": { "type": "integer" }, "targetsIdentified": { "type": "integer" }, "compoundsCurated": { "type": "integer" }, "patentsAnalysed": { "type": "integer" } } },
      "QuickAction": { "type": "object", "properties": { "label": { "type": "string" }, "module": { "type": "string" }, "prefillQuery": { "type": "string" } } },
      "PipelineProgress": { "type": "object", "properties": { "target": { "type": "string" }, "pipelineType": { "type": "string" }, "progressPercent": { "type": "integer" } } },
      "PipelineDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/PipelineProgress" },
          { "type": "object", "properties": { "id": { "type": "string" }, "status": { "type": "string" }, "projectId": { "type": "string", "nullable": true }, "jobId": { "type": "string", "nullable": true }, "updatedAt": { "type": "string", "format": "date-time" } } }
        ]
      },
...
      "TargetProfileRequest": {
        "type": "object",
        "required": ["target"],
        "properties": {
          "target": { "type": "string" }, "sourceSessionId": { "type": "string", "nullable": true },
          "disease": { "type": "string", "nullable": true }, "weights": { "type": "object", "additionalProperties": { "type": "number" } }
        }
      },
      "CurateCompoundsRequest": {
        "type": "object",
        "required": ["target"],
        "properties": {
          "target": { "type": "string" }, "compound": { "type": "string", "nullable": true },
          "numResults": { "type": "integer", "minimum": 15, "maximum": 50, "default": 20 },
          "disease": { "type": "string", "nullable": true }, "weights": { "type": "object", "additionalProperties": { "type": "number" } }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth (legacy scaffold)"], "summary": "Login (STUB - unimplemented, returns nothing)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } } } },
        "responses": { "200": { "description": "OK (stub)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserOut" } } } } }
      }
    },
...
    "/api/v1/drug-curation/curatex/profile": {
      "post": {
        "tags": ["3. Drug Curation"], "summary": "CurateX Tool 1 - profile",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileRequest" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/auth/login": {
      "post": {
        "tags": ["DRP Auth & Onboarding"], "summary": "Login",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } } }
      }
    },
    "/v1/auth/logout": { "post": { "tags": ["DRP Auth & Onboarding"], "summary": "Logout", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } } } },
    "/v1/auth/refresh": {
      "post": {
        "tags": ["DRP Auth & Onboarding"], "summary": "Refresh token",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshResponse" } } } } }
      }
    },
    "/v1/auth/forgot-password": {
      "post": {
        "tags": ["DRP Auth & Onboarding"], "summary": "Forgot password (always 200)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/v1/users/me": { "get": { "tags": ["DRP Auth & Onboarding"], "summary": "Current user profile", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DrpUser" } } } } } } },
    "/v1/users/me/onboarding-status": { "get": { "tags": ["DRP Auth & Onboarding"], "summary": "Onboarding status", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OnboardingStatus" } } } } } } },
    "/v1/users/me/research-focus": {
      "post": {
        "tags": ["DRP Auth & Onboarding"], "summary": "Save research focus (therapeutic areas)", "security": [{ "drpToken": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResearchFocusRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/v1/users/me/onboarding/complete": { "post": { "tags": ["DRP Auth & Onboarding"], "summary": "Complete onboarding", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } } } },
    "/v1/therapeutic-areas": { "get": { "tags": ["DRP Auth & Onboarding"], "summary": "List therapeutic areas", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } } } } } },
    "/v1/dashboard/summary": { "get": { "tags": ["DRP Dashboard"], "summary": "Dashboard summary", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DashboardSummary" } } } } } } },
    "/v1/dashboard/quick-actions": { "get": { "tags": ["DRP Dashboard"], "summary": "Quick actions", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/QuickAction" } } } } } } } },
    "/v1/dashboard/pipelines": { "get": { "tags": ["DRP Dashboard"], "summary": "Pipeline progress", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PipelineProgress" } } } } } } } },
    "/v1/dashboard/pipelines/manage": { "get": { "tags": ["DRP Dashboard"], "summary": "Manage pipelines (detailed)", "security": [{ "drpToken": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PipelineDetail" } } } } } } } },
...
    "/v1/agents/curatex/target-profile": {
      "post": {
        "tags": ["DRP Agents - CuraTeX"], "summary": "CuraTeX - Target profile (creates job)", "security": [{ "drpToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TargetProfileRequest" } } } },
        "responses": { "202": { "description": "Job accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobAccepted" } } } } }
      }
    },
    "/v1/agents/curatex/compounds": {
      "post": {
        "tags": ["DRP Agents - CuraTeX"], "summary": "CuraTeX - Curate compounds (creates job)", "security": [{ "drpToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CurateCompoundsRequest" } } } },
        "responses": { "202": { "description": "Job accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobAccepted" } } } } }
      }
    },
...
  }
}